+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Compatibility with hypothesis >= 6.131.4. [#2797](https://github.com/schemathesis/schemathesis/issues/2797)
 - Cover more nested `enum` values during the coverage phase.
 - Avoid extremely slow generation of large strings combined with `patterns` during the coverage phase.
+- Generate objects during the coverage phase even if `"type": "object"` is not present.
 
 ## [4.0.0-alpha.8](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.7...v4.0.0-alpha.8) - 2025-04-05